+++ CHANGELOG.rst
@@ -12,6 +12,10 @@ Performance
 ~~~~~~~~~~~
 * add banded implementation of LCS / Indel. This improves the runtime from ``O((|s1|/64) * |s2|)`` to ``O((score_cutoff/64) * |s2|)``
 
+Changed
+~~~~~~~
+* upgrade to ``Cython==3.0.7``
+
 [3.5.2] - 2023-11-02
 ^^^^^^^^^^^^^^^^^^^^
 Fixed